# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ElevateKe Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trade/place:
    post:
      operationId: placeTrade
      tags: [trade]
      summary: Place a trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceTradeBody"
      responses:
        "201":
          description: Trade placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trade/cashout:
    post:
      operationId: cashoutTrade
      tags: [trade]
      summary: Cashout active trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CashoutTradeBody"
      responses:
        "200":
          description: Cashout result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResult"
        "400":
          description: No active trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PlaceTradeBody:
      type: object
      properties:
        amount:
          type: number
        multiplier:
          type: integer
        durationMins:
          type: integer
        direction:
          type: string
          enum: [up, down]
      required: [amount, multiplier, durationMins, direction]
    CashoutTradeBody:
      type: object
      properties:
        tradeId:
          type: integer
      required: [tradeId]
    Trade:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        multiplier:
          type: integer
        durationMins:
          type: integer
        direction:
          type: string
        result:
          type: ["string", "null"]
        profitLoss:
          type: ["number", "null"]
        status:
          type: string
        startedAt:
          type: string
        endedAt:
          type: ["string", "null"]
      required: [id, amount, multiplier, durationMins, direction, status, startedAt]
    TradeResult:
      type: object
      properties:
        trade:
          $ref: "#/components/schemas/Trade"
        profitLoss:
          type: number
        newBalance:
          type: number
        message:
          type: string
      required: [trade, profitLoss, newBalance, message]
...